% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{add_vaf}
\alias{add_vaf}
\title{Adding the VAF dataset to the current object.}
\usage{
add_vaf(x, vaf_df)
}
\arguments{
\item{x}{a mvnmm object.}

\item{vaf_df}{a matrix-like dataset, containing the VAF of the mutations.}
}
\value{
a mvnmm object.
}
\description{

}
