% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov.df.example}
\alias{cov.df.example}
\title{Example coverage data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 428 rows and 4 columns.
}
\usage{
data(cov.df.example)
}
\description{
Example coverage data
}
\examples{
data(cov.df.example)
head(cov.df.example)
}
\keyword{datasets}
