% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mullerplot.R
\name{estimate_n_pops}
\alias{estimate_n_pops}
\title{Function implemented to estimate the real number of clones in each cluster.}
\usage{
estimate_n_pops(x, highlight = c())
}
\arguments{
\item{x}{mvnmm object}

\item{highlight}{clusters to show}
}
\value{
named array reporting for each cluster the estimated true number of populations.
}
\description{
Function implemented to estimate the real number of clones in each cluster.
}
