% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Filters the input dataset.}
\usage{
filter_dataset(
  dataset,
  columns,
  IS,
  min_cov = 50,
  min_ccf = 0.05,
  k_interval = c(5, 30),
  metric = "calinski_harabasz_score",
  random_state = 25
)
}
\arguments{
\item{dataset}{Input dataset. It has to be of shape \code{NxT}, where \code{N} is the number of
observations and \code{T} is the dimensionality of the dataset.}

\item{columns}{List of columns to be selected from the dataset. If not specified,
all columns will be considered.}

\item{min_cov}{add}

\item{min_ccf}{add}

\item{k_interval}{add}

\item{metric}{add}

\item{random_state}{add}

\item{IS_values}{List of Insertion Sites identifier. It must be a list of length \code{N}, hence
for each observation a value is needed. If not specified, a general identifier \code{IS.1,...IS.N}
will be generated.}
}
\value{
a dataset of the same shape as the input one, with filtered observations.
}
\description{

}
