% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{Creates an object of class \code{mvnmm}.}
\usage{
fit(
  cov.df,
  vaf.df = NULL,
  infer_phylogenies = TRUE,
  hyperparameters = list(),
  k_interval = c(10, 30),
  n_runs = 3,
  steps = 500,
  lr = 0.005,
  p = 0.01,
  convergence = TRUE,
  covariance = "diag",
  min_frac = 0,
  show_progr = FALSE,
  store_grads = TRUE,
  store_losses = TRUE,
  random_state = 25
)
}
\arguments{
\item{cov.df}{Input coverage dataset. It must have at least the columns \code{coverage}, \code{timepoints},
\code{lineage}, \code{IS}, with the coverage values, timepoint, lineage and IS, respectively.}

\item{vaf.df}{Input VAF dataset. It must have at least the columns \code{alt}, \code{dp}, \code{vaf},
\code{timepoints}, \code{lineage}, \code{IS}, \code{mutation}, with the number of reads for the mutated allele,
overall depth, vaf values, timepoint, lineage, IS and mutation, respectively.}

\item{k_interval}{Interval of K values to test.}

\item{n_runs}{Number of runs to perform for each K.}

\item{steps}{Maximum number of steps for the inference.}

\item{lr}{Learning rate used in the inference.}

\item{covariance}{Covariance type for the Multivariate Gaussian.}

\item{random_state}{Value of the seed.}
}
\value{
a \code{mvnmm} object, containing the input dataset, annotated with IS_values, N, K, T
specific of the dataset, the input IS and column names, a list params that will contain the
inferred parameters, the python object
}
\description{
Function to fit the input data.
}
