% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_model.R
\name{fit_growth_rates}
\alias{fit_growth_rates}
\title{Infer growth rates for each clone and subclone.}
\usage{
fit_growth_rates(
  x,
  steps = 500,
  highlight = list(),
  timepoints_to_int = list(),
  growth = "",
  force = T
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{steps}{add}

\item{highlight}{add}

\item{timepoints_to_int}{add}

\item{growth}{add}

\item{force}{add}
}
\value{

}
\description{
Infer growth rates for each clone and subclone.
}
