% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations_clust.R
\name{fit_mutations}
\alias{fit_mutations}
\title{Fit the mutations clustering}
\usage{
fit_mutations(
  x,
  vaf.df = NULL,
  infer_phylo = TRUE,
  min_frac = 0,
  max_IS = NULL,
  highlight = list(),
  lineages = c(),
  label = ""
)
}
\arguments{
\item{x}{add}

\item{vaf.df}{add}

\item{infer_phylo}{add}

\item{min_frac}{add}

\item{max_IS}{add}

\item{highlight}{add}

\item{lineages}{add}

\item{label}{add}
}
\value{

}
\description{
Fit the mutations clustering
}
