% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenies_inference.R
\name{fit_phylogenies}
\alias{fit_phylogenies}
\title{Fit the phylogenetic trees}
\usage{
fit_phylogenies(
  x,
  vaf.df = NULL,
  min_frac = 0,
  highlight = list(),
  do_filter = FALSE,
  label = "",
  fit_muts = FALSE,
  lineages = c()
)
}
\arguments{
\item{x}{add}

\item{vaf.df}{add}

\item{min_frac}{add}

\item{highlight}{add}

\item{do_filter}{add}

\item{label}{add}

\item{lineages}{add}

\item{fit_viber}{add}
}
\value{

}
\description{
Fit the phylogenetic trees
}
