% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_cov_dataframe}
\alias{get_cov_dataframe}
\title{Retrieve the coverage dataframe.}
\usage{
get_cov_dataframe(x, verbose = T)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
the coverage dataset used to fit the model.
}
\description{
Function to retrieve the coverage dataframe used to initialize the object.
}
\examples{
if (FALSE) get_cov_dataframe(x)

}
