% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_covariance_Sigma}
\alias{get_covariance_Sigma}
\title{Extract the estimated covariance matrices.}
\usage{
get_covariance_Sigma(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
list of the estimated covariance matrices.
}
\description{
Returns a list with \code{K} dataframes, each of dimension \code{TxT}, corresponding
to the covariance matrices estimated for each clone \code{k}
}
\examples{
if (FALSE) get_covariance_Sigma(x)

}
