% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_dataframe}
\alias{get_dataframe}
\title{Retrieve the object dataframe.}
\usage{
get_dataframe(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
the dataset used to fit the model.
}
\description{

}
