% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Extract the model dimensions.}
\usage{
get_dimensions(x, verbose = T)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
vector of model dimensions.
}
\description{
Returns a vector with the dimensions of the model.
}
\examples{
if (FALSE) get_dimensions(x)

}
