% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_growth_rates}
\alias{get_growth_rates}
\title{Get infered growth rates dataframe.}
\usage{
get_growth_rates(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get infered growth rates dataframe.
}
