% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_ISs}
\alias{get_ISs}
\title{Get the number of ISs per cluster.}
\usage{
get_ISs(x, highlight = c())
}
\arguments{
\item{x}{the fitted object}

\item{highlight}{the clusters to retrieve}
}
\value{
an array with as names the clusters in \code{highlight} and as values the number of ISs assigned
to each cluster
}
\description{
Get the number of ISs per cluster.
}
