% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_labels}
\alias{get_labels}
\title{Extract the observations labels.}
\usage{
get_labels(x, init = F)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
list of observations labels.
}
\description{
Returns a list with \code{N} elements, corresponding to the labels for each
observation.
}
\examples{
if (FALSE) get_labels(x)

}
