% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_lineages}
\alias{get_lineages}
\title{Extract the data lineages.}
\usage{
get_lineages(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
vector of data lineages.
}
\description{
Returns a vector with the lineages of the input data.
}
\examples{
if (FALSE) get_lineages(x)

}
