% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_sigma}
\alias{get_sigma}
\title{Extract the estimated variance parameters.}
\usage{
get_sigma(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
dataframe of variance parameters.
}
\description{
Returns a dataframe \code{KxT} with the estimated variance paramaters \code{sigma_kt} per
clone \code{k} and dimension \code{t}.
}
\examples{
get_sigma(x)

}
