% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_timepoints}
\alias{get_timepoints}
\title{Extract the data timepoints.}
\usage{
get_timepoints(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
vector of data timepoints.
}
\description{
Returns a vector with the timepoints of the input data.
}
\examples{
get_timepoints(x)

}
