% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{get_unique_labels}
\alias{get_unique_labels}
\title{Functions to access elements of a mvnmm object.}
\usage{
get_unique_labels(obj)
}
\description{
Functions to access elements of a mvnmm object.
}
