% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_unique_muts_labels}
\alias{get_unique_muts_labels}
\title{Retrieve the list of unique labels of mutation clusters.}
\usage{
get_unique_muts_labels(x, clusters = c(), verbose = F)
}
\arguments{
\item{x}{a mvnmm object.}

\item{clusters}{a vector-like variable, with the identifiers of the clones we want
to retrieve the subclone labels from. If empty, all the labels will be returned.}
}
\value{
vector of mutations labels.
}
\description{
Function to retrieve the list of unique labels of mutations clusters,
of the form \code{C_c1.Cm1}, where \code{c1} is the clone identifier and \code{m1}
is the subclone identifier.
}
\examples{
if(FALSE) get_unique_muts_labels(x, c("C_0"))

}
