% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_vaf_dataframe}
\alias{get_vaf_dataframe}
\title{Retrieve the mutations dataframe.}
\usage{
get_vaf_dataframe(x, label = "")
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
the mutations dataset.
}
\description{
Function to retrieve the mutations dataframe used to initialize the object.
}
\examples{
if (FALSE) get_vaf_dataframe(x)

}
