% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_viber_clusters}
\alias{get_viber_clusters}
\title{Retrieve the unique labels of mutation clusters, given the coverage labels.}
\usage{
get_viber_clusters(x, clusters)
}
\arguments{
\item{x}{a mvnmm object.}

\item{clusters}{a vector-like variable, with the coverage labels.}
}
\value{
vector of mutations labels.
}
\description{

}
