% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_weights}
\alias{get_weights}
\title{Extract the estimated mixing proportions.}
\usage{
get_weights(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
list of estimated mixing proportions.
}
\description{
Returns a list of dimension \code{K} with the estimated mixing proportions.
}
\examples{
if (FALSE) get_weights(x)

}
