% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_z_probs}
\alias{get_z_probs}
\title{Extract the estimated posterior probabilities.}
\usage{
get_z_probs(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
dataframe of posterior distributions.
}
\description{
Returns a dataframe of shape \code{NxK} with the posterior distribution \code{p(k|n)}
for each observation \code{n} to belong to cluster \code{k}.
}
\examples{
get_z_probs(x)

}
