% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{mixture_model}
\alias{mixture_model}
\title{Creates an object of class mvnmm.}
\usage{
mixture_model(K, dataset, lineages, columns = list(), IS_values = list())
}
\arguments{
\item{K}{Input maximum number of clusters}

\item{dataset}{Input dataset. It has to be of shape NxT, where N is the number of
observations and T is the dimensionality of the dataset.}

\item{columns}{List of columns to be selected from the dataset. If not specified,
all columns will be considered.}

\item{IS_values}{List of Insertion Sites identifier. It must be a list of length N, hence
for each observation a value is needed. If not specified, a general identifier IS.1,...IS.N
will be generated.}
}
\value{
a LineaGT object, containing the input dataset, annotated with IS_values,
N, K, T specific of the dataset, the input IS and column names, a list params that
will contain the inferred parameters, the python object
}
\description{
Add
}
