% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_differentiation_tree.R
\name{plot_differentiation_tree}
\alias{plot_differentiation_tree}
\title{Visualize the number of subclones on the differentiatio tree}
\usage{
plot_differentiation_tree(x, label = "", label.on.node = T, cex = 1)
}
\arguments{
\item{x}{add}

\item{label}{add}

\item{label.on.node}{add}

\item{cex}{add}
}
\value{

}
\description{
Visualize the number of subclones on the differentiatio tree
}
