% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_clones.R
\name{plot_exp_rate}
\alias{plot_exp_rate}
\title{Exponential rate}
\usage{
plot_exp_rate(
  x,
  highlight = c(),
  min_frac = 0,
  mutations = F,
  label = "",
  timepoints_to_int = list(init = 0, early = 60, mid = 140, late = 280)
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{highlight}{a vector of clusters IDs to highlight in the plot.}

\item{min_frac}{min_frac numeric value in \code{[0,1]} representing the minimum abundance to highlight a clone.}

\item{mutations}{Boolean. If set to \code{TRUE}, the growth will be visualize for each cluster of mutations.}

\item{label}{a character corresponding to the label of the run to visualize.}

\item{timepoints_to_int}{a list to map each \code{timepoint} value to an integer.}

\item{facet}{Boolean. If set to \code{TRUE}, the plot will be faceted against the clusters.}
}
\description{
Function to visualize the growth coefficients for each clone and lineage.
}
\examples{
if (FALSE) plot_exp_rate(x)

}
