% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diagn.R
\name{plot_gradient_norms}
\alias{plot_gradient_norms}
\title{Function to plot the gradients norms.}
\usage{
plot_gradient_norms(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\description{
The gradient norms of the parameters per iteration, computed for each input \code{K}
used during model selection.
}
