% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_growth.R
\name{plot_growth_rates}
\alias{plot_growth_rates}
\title{Visualize the infered growth rates.}
\usage{
plot_growth_rates(
  x,
  highlight = c(),
  min_frac = 0,
  mutations = F,
  timepoints_to_int = list(),
  fit = F
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{highlight}{a vector of clusters IDs to highlight in the plot.}

\item{min_frac}{min_frac numeric value in \code{[0,1]} representing the minimum abundance to highlight a clone.}

\item{mutations}{Boolean. If set to \code{TRUE}, the growth will be visualize for each cluster of mutations.}

\item{timepoints_to_int}{a list to map each \code{timepoint} value to an integer.}

\item{fit}{add}
}
\description{
Function to visualize the growth coefficients for each clone and lineage.
}
\examples{
if (FALSE) plot_growth_rates(x)

}
