% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diagn.R
\name{plot_IC}
\alias{plot_IC}
\title{Function to plot the Information Criteria computed during model selection.}
\usage{
plot_IC(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\description{
Function to plot the Information Criteria computed during model selection.
}
