% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diagn.R
\name{plot_losses}
\alias{plot_losses}
\title{Function to plot the training losses.}
\usage{
plot_losses(x, train = FALSE)
}
\arguments{
\item{x}{a mvnmm object.}

\item{train}{Boolean. If set to \code{TRUE}, the losses computed in the model selection are visualized.}
}
\description{
Function to plot the training losses.
}
