% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_mixture_weights}
\alias{plot_mixture_weights}
\title{Barplot of the per-cluster mixture weights and number of ISs.}
\usage{
plot_mixture_weights(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\description{
add
}
\examples{
if (FALSE) plot_mixture_weights(x)

}
