% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_trees.R
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Clonal evolution trees}
\usage{
plot_phylogeny(
  x,
  score_diff = 1,
  show_best = 1,
  min_frac = 0,
  highlight = c(),
  label = ""
)
}
\arguments{
\item{x}{add}

\item{score_diff}{add}

\item{show_best}{add}

\item{min_frac}{add}

\item{highlight}{add}

\item{label}{add}
}
\value{

}
\description{
Clonal evolution trees
}
