% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_scatter_density}
\alias{plot_scatter_density}
\title{Plot scatter and density}
\usage{
plot_scatter_density(obj, plot_density = T, facet = FALSE, highlight = c())
}
\arguments{
\item{obj}{a mvnmm object.}

\item{plot_density}{a Boolean. If set to FALSE, the Gaussian density will not be displayed.}

\item{facet}{a Boolean. If set to TRUE, the plot will be faceted on the labels.}
}
\value{
a list of plots, one for each timepoints combination.
}
\description{
Function to plot the scatterplots of the coverage, one timepoint against the other, together with the Gaussian density.
}
