% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_vaf.R
\name{plot_vaf}
\alias{plot_vaf}
\title{VAF scatterplot}
\usage{
plot_vaf(obj, min_ccf = 0, highlight = c())
}
\arguments{
\item{obj}{a mvnmm object}

\item{min_ccf}{value in [0,1] to show only the clusters with a minimum CCF of min_ccf in at least one timepoint}

\item{highlight}{a list of labels ID to show. All the clusters in the list are shown, even if the CCF is lower than min_ccf}
}
\description{
Function to plot the VAFs of the mutations one timepoint against the other
}
\examples{
obj = run_viber(obj, lineaGT::homo_ltr.vaf, min_ccf=0.07)
plot_vaf(obj)
}
