% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_vaf.R
\name{plot_vaf_time}
\alias{plot_vaf_time}
\title{VAF over time}
\usage{
plot_vaf_time(x, min_frac = 0, highlight = c(), timepoints_to_int = list())
}
\arguments{
\item{x}{a \code{mvnmm} object.}

\item{min_frac}{value in \code{[0,1]} representing the minimum abundance to show the clusters.}

\item{highlight}{a list of labels ID to show.}

\item{timepoints_to_int}{a list to map each \code{timepoint} value to an integer.}
}
\value{
a \code{ggplot} object.
}
\description{
Function to plot the VAFs of the mutations along time
}
\examples{
if (FALSE) plot_vaf_time(x, min_frac=0.1)

}
