% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{print.mvnmm}
\alias{print.mvnmm}
\title{Print method}
\usage{
\method{print}{mvnmm}(x, ...)
}
\arguments{
\item{x}{An object of class mvnmm}

\item{...}{Default extra paramaters}
}
\value{

}
\description{
add
}
