% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.example}
\alias{x.example}
\title{Example mutation data}
\format{
An object of class \code{mvnmm} of length 19.
}
\usage{
data(vaf.df.example)
}
\description{
Example mutation data
}
\examples{
data(vaf.df.example)
head(vaf.df.example)
}
\keyword{datasets}
