% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filtering.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Filters the input dataset.}
\usage{
filter_dataset(
  cov.df,
  min_cov = 5,
  min_frac = 0.05,
  k_interval = c(10, 20),
  metric = "calinski_harabasz_score",
  seed = 5
)
}
\arguments{
\item{cov.df}{Input coverage dataset. It must have at least the columns \code{coverage}, \code{timepoints},
\code{lineage}, \code{IS}, with the coverage values, timepoint, lineage and IS, respectively.}

\item{min_cov}{add}

\item{min_frac}{add}

\item{k_interval}{add}

\item{metric}{add}

\item{seed}{add}
}
\value{
a dataset of the same shape as the input one, with filtered observations.
}
\description{

}
