% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{fit}
\alias{fit}
\title{Creates an object of class \code{mvnmm}.}
\usage{
fit(
  dataset,
  lineages,
  columns = list(),
  IS_values = list(),
  k_interval = c(10, 30),
  n_runs = 3,
  steps = 500,
  lr = 0.005,
  covariance = "diag",
  random_state = 25
)
}
\arguments{
\item{dataset}{Input dataset. It has to be of shape \code{NxT}, where \code{N} is the number of
observations and \code{T} is the dimensionality of the dataset.}

\item{lineages}{List of the lineages into consideration.}

\item{columns}{List of columns to be selected from the dataset. If not specified,
all columns will be considered.}

\item{IS_values}{List of Insertion Sites identifier. It must be a list of length \code{N}, hence
for each observation a value is needed. If not specified, a general identifier \code{IS.1,...IS.N}
will be generated.}

\item{k_interval}{Interval of K values to test.}

\item{n_runs}{Number of runs to perform for each K.}

\item{steps}{Maximum number of steps for the inference.}

\item{lr}{Learning rate used in the inference.}

\item{covariance}{Covariance type for the Multivariate Gaussian.}

\item{random_state}{Value of the seed.}
}
\value{
a \code{mvnmm} object, containing the input dataset, annotated with IS_values, N, K, T
specific of the dataset, the input IS and column names, a list params that will contain the
inferred parameters, the python object
}
\description{
Function to fit the input data.
}
