% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations_clust.R
\name{fit_mutations}
\alias{fit_mutations}
\title{Fit the mutations clustering}
\usage{
fit_mutations(
  x,
  vaf.df = NULL,
  infer_phylo = TRUE,
  min_frac = 0,
  max_IS = NULL,
  highlight = list()
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{vaf.df}{dataframe with mutations data.}

\item{infer_phylo}{a Boolean indicating whether to infer also the phylogenetic evolution per cluster of ISs.}

\item{min_frac}{add}

\item{max_IS}{add}

\item{highlight}{add}
}
\value{
A \code{mvnmm} object with the additional list \code{x.muts} containing the estimated subclones from somatic mutations.
}
\description{
Fit the mutations clustering
}
