% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_params.R
\name{get_unique_labels}
\alias{get_unique_labels}
\title{Extract the list of unique observations labels.}
\usage{
get_unique_labels(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
list of unique labels.
}
\description{
Returns a list with \code{K} elements, corresponding to the unique labels.
}
\examples{
if (FALSE) get_unique_labels(x)

}
