% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_dataframe.R
\name{get_vaf_dataframe}
\alias{get_vaf_dataframe}
\title{Retrieve the object VAF dataframe.}
\usage{
get_vaf_dataframe(x)
}
\arguments{
\item{x}{a mvnmm object.}
}
\value{
the VAF dataset.
}
\description{

}
