% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_plots.R
\name{plot.mvnmm}
\alias{plot.mvnmm}
\title{2D scatter and density plot}
\usage{
\method{plot}{mvnmm}(x, ...)
}
\arguments{
\item{x}{An object of class mvnmm}

\item{...}{Default extra paramaters}
}
\value{

}
\description{
2D scatter and density plot for each pair of dimensions
}
