% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_differentiation_tree.R
\name{plot_differentiation_tree}
\alias{plot_differentiation_tree}
\title{Visualize the number of subclones on the differentiatio tree}
\usage{
plot_differentiation_tree(
  x,
  edges = differentiation_tree(),
  highlight = c(),
  single_tree = T,
  wrap = T,
  timepoints = c()
)
}
\arguments{
\item{x}{add}

\item{edges}{add}

\item{highlight}{add}

\item{single_tree}{add}

\item{wrap}{add}
}
\description{
Visualize the number of subclones on the differentiatio tree
}
