% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_marginal}
\alias{plot_marginal}
\title{Histogram of the marginal distribution of each dimension}
\usage{
plot_marginal(
  x,
  min_frac = 0,
  highlight = c(),
  binwidth = 5,
  show_mean = F,
  timepoints_to_int = list()
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{highlight}{a vector of clusters IDs to highlight in the plot.}

\item{binwidth}{numeric value representing the histogram binwidth.}
}
\description{
Function to plot the marginal distribution of the coverage, for each timepoint, colored by cluster.
}
\examples{
if (FALSE) plot_marginal(x)

}
