% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_clones.R
\name{plot_mullerplot}
\alias{plot_mullerplot}
\title{Muller plot}
\usage{
plot_mullerplot(
  obj,
  which = "ccf",
  highlight = c(),
  min_ccf = 0,
  legend.pos = "right",
  wrap = F
)
}
\description{

}
