% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mullerplot.R
\name{plot_mullerplot}
\alias{plot_mullerplot}
\title{Muller plot}
\usage{
plot_mullerplot(
  x,
  which = "frac",
  highlight = c(),
  min_frac = 0,
  timepoints_to_int = c(),
  mutations = F,
  single_clone = T,
  tree_score = 1,
  legend.pos = "right"
)
}
\arguments{
\item{x}{a mvnmm object.}

\item{which}{string among \code{"frac","pop","fitness"} determining whether to plot the coverage
normalized in \code{[0,1]}, as absolute clone abundance, or with each clone colored by the growth rate,
computed assuming a exponential growth.}

\item{highlight}{a vector of clusters IDs to highlight in the plot.}

\item{min_frac}{min_frac numeric value in \code{[0,1]} representing the minimum abundance to highlight a clone.}

\item{timepoints_to_int}{a list to map each \code{timepoint} value to an integer.}

\item{mutations}{Boolean. If set to \code{TRUE}, also the clusters of mutations will be visualized.}

\item{single_clone}{Boolean. If \code{mutations} and \code{single_clone} are set to \code{TRUE}, only the clones
reported in \code{highlight} and the respective subclones will be visualised.}

\item{tree_score}{add}

\item{legend.pos}{add}
}
\description{
Function to visualize the mullerplot for the fitted object.
}
\examples{
if (FALSE) plot_mullerplot(x, wrap=T)

}
