% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_phylogenies.R
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Clonal evolution trees}
\usage{
plot_phylogeny(x, show_best = 1, min_frac = 0, highlight = c())
}
\arguments{
\item{x}{a \code{mvnmm} object.}

\item{show_best}{the number of trees to visualize based on the computed score.}

\item{min_frac}{value in \code{[0,1]} representing the minimum abundance to show the clusters.}

\item{highlight}{a list of labels ID to show.}
}
\description{
Clonal evolution trees
}
