% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_scatter_density}
\alias{plot_scatter_density}
\title{2D scatterplot and density}
\usage{
plot_scatter_density(x, plot_density = T, highlight = c(), min_frac = 0)
}
\arguments{
\item{x}{a mvnmm object.}

\item{plot_density}{a Boolean. If set to FALSE, the Gaussian density will not be displayed.}

\item{highlight}{a vector of clusters IDs to highlight in the plot.}

\item{facet}{a Boolean. If set to TRUE, the plot will be faceted on the labels.}
}
\value{
a list of plots, one for each timepoints combination.
}
\description{
Function to plot the scatterplots of the coverage, one timepoint against the other, together with the Gaussian density.
}
\examples{
if (FALSE) plots = plot_scatter_density(x)

}
