% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_vaf.R
\name{plot_vaf}
\alias{plot_vaf}
\title{VAF scatterplot}
\usage{
plot_vaf(x, min_frac = 0, highlight = c())
}
\arguments{
\item{x}{a \code{mvnmm} object.}

\item{highlight}{a list of labels ID to show. All the clusters in the list are shown,
even if the CCF is lower than \code{min_frac}.}

\item{min_fracc}{value in \code{[0,1]} to show only the clusters with a minimum CCF of
\code{min_frac} in at least one timepoint.}
}
\value{
list of VAF scatterplots.
}
\description{
Function to plot the VAFs of the mutations one timepoint against the other
}
