% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_vaf.R
\name{plot_vaf}
\alias{plot_vaf}
\title{VAF 2D scatterplot}
\usage{
plot_vaf(x, min_frac = 0, highlight = c(), wrap = T)
}
\arguments{
\item{x}{a \code{mvnmm} object.}

\item{min_frac}{value in \code{[0,1]} representing the minimum abundance to show the clusters.}

\item{highlight}{a list of labels ID to show.}

\item{wrap}{a Boolean to wrap the scatterplots of multiple lineages in a unique plot.}
}
\value{
list of VAF scatterplots.
}
\description{
Function to plot the VAFs of the mutations one timepoint against the other
}
\examples{
if (FALSE) plot_vaf(x, min_frac=0.1)

}
