% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations_clust.R
\name{run_viber}
\alias{run_viber}
\title{Functions to run VIBER for mutations clustering}
\usage{
run_viber(obj, min_ccf = 0, highlight = list())
}
\arguments{
\item{obj}{mvnmm object.}

\item{vaf_df}{a VAF dataframe, with required columns mutation,dp,ref,alt of depth, reference and alternative reads}
}
\value{
mvnmm object with added the VAF dataframe annotated with the found clusters and the VIBER fit for each cluster
}
\description{
add
}
